Custom Date properties can edit times again.
Fixes ZEN-27345.

Also adds the Zenoss.form.field.DateTime and Zenoss.date.Moment
classes for handling datetimes and timezones.

+++ Products/Zuul/routers/zep.py
@@ -233,7 +233,7 @@ class EventsRouter(DirectRouter):
             values = []
             splitter = ' TO ' if ' TO ' in value else '/'
             for t in value.split(splitter):
-                values.append(int(t))
+                values.append(float(t))
             return values
         except ValueError:
             log.warning("Invalid timestamp: %s", value)
